Add files via upload
The pdf document contains my jupyter notebook codes.
I lost the original ipynb files at the point of doing this due to some tech itch.
I will be recreating the notebbok in a couple of hours.
Please kindly treat the pdf document as my notebook submission. I had to do this because of time constraints.